requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
google-adk==0.5.0
biopython>=1.84
pypdf>=4.3.1
sentence-transformers>=3.0.1
httpx>=0.27.0
python-dotenv>=1.0.1
pydantic>=2.8.0
google-cloud-storage>=2.18.0
google-cloud-aiplatform>=1.87.0
aiofiles>=23.2.1
numpy>=1.24.0
